rename model to models
