+++ models/document.go
@@ -1,4 +1,4 @@
-package model
+package models
 
 import "github.com/jinzhu/gorm"
